Move timestamp to its own module
